+++ test/database_test_helper.rb
@@ -23,9 +23,11 @@ ActiveRecord::Migration.verbose = false
 
 require_relative "../db/migrate/20260805000000_create_solid_objects_tables"
 require_relative "../db/migrate/20260806000000_add_state_revision_to_solid_objects_instances"
+require_relative "../db/migrate/20260813000000_rename_message_dispatch_columns"
 
 CreateSolidObjectsTables.new.migrate(:up)
 AddStateRevisionToSolidObjectsInstances.new.migrate(:up)
+RenameMessageDispatchColumns.new.migrate(:up)
 
 ActiveRecord::Base.connection.create_table(:solid_objects_test_domain_records) do |table|
   table.string :name, null: false